backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.0.2
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
requests==2.31.0
numpy==1.24.3
pillow==10.1.0
sentinelhub==3.9.0
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2